+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Changelog
 
+- External reference is forwarded to a wrapped component by `withPopupMenu`.
+
 ### 0.10.0 (2020-05-19)
 
 - `withPopupMenu` higher order component is added.